product/requirements.txt
fastapi==0.98.0
pydantic==1.10.7
alembic==1.10.3
python-dotenv==0.19.2
black==23.7.0
pandas==2.0.3
anyio==3.7.1
boltons==23.0.0
boto3==1.28.35
botocore==1.31.35
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
docker==6.1.3
furl==2.1.3
googleapis-common-protos==1.60.0
grpcio==1.57.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==2.10
jmespath==1.0.1
orderedmultidict==1.0.1
packaging==23.1
protobuf==4.24.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.2
setuptools-scm==7.1.0
six==1.16.0
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==1.26.16
websocket-client==1.6.2
yandexcloud==0.228.0
yappa==0.4.28
json2yaml
